DATA-1867: move election_api_utils back into include/
Astro's DAG-only deploy doesn't ship include/, but full image deploys
do — and the include/ layout is conventional. Adopting full deploys for
this DAG, per the team's standard pattern.
